# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UAT Test Case Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /use-cases/{useCaseId}/test-cases:
    get:
      operationId: listTestCases
      tags: [test-cases]
      summary: List test cases for a use case
      parameters:
        - name: useCaseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of test cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestCase"
    post:
      operationId: createTestCase
      tags: [test-cases]
      summary: Create a test case
      parameters:
        - name: useCaseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestCaseBody"
      responses:
        "201":
          description: Created test case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
  /test-cases/{testCaseId}:
    put:
      operationId: updateTestCase
      tags: [test-cases]
      summary: Update a test case
      parameters:
        - name: testCaseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestCaseBody"
      responses:
        "200":
          description: Updated test case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
    delete:
      operationId: deleteTestCase
      tags: [test-cases]
      summary: Delete a test case
      parameters:
        - name: testCaseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Test Steps ────────────────────────────────────────────────────────────
...
  /test-runs/{testRunId}/discussions:
    post:
      operationId: createDiscussion
      tags: [team-discussions]
      summary: Start a team discussion for a test run
      parameters:
        - name: testRunId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiscussionBody"
      responses:
        "201":
          description: Discussion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDiscussion"
  /discussions/{discussionId}:
    get:
      operationId: getDiscussion
      tags: [team-discussions]
      summary: Get a team discussion with participants
      parameters:
        - name: discussionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discussion detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDiscussion"
    patch:
      operationId: endDiscussion
      tags: [team-discussions]
      summary: End a team discussion (Test Lead/Admin)
      parameters:
        - name: discussionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discussion ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDiscussion"
  /discussions/{discussionId}/participants:
    post:
      operationId: addParticipant
      tags: [team-discussions]
      summary: Add a participant to a discussion
      parameters:
        - name: discussionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddParticipantBody"
      responses:
        "201":
          description: Participant added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDiscussion"
  /discussions/{discussionId}/participants/{userId}:
    delete:
      operationId: removeParticipant
      tags: [team-discussions]
      summary: Remove a participant from a discussion
      parameters:
        - name: discussionId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Participant removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDiscussion"
  /discussions/{discussionId}/defects/{defectId}:
    get:
      operationId: getDiscussionDefectDrilldown
      tags: [team-discussions]
      summary: Get full defect drill-down for a discussion
      parameters:
        - name: discussionId
          in: path
          required: true
          schema:
            type: integer
        - name: defectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Defect drill-down data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Defect"
components:
  schemas:
...
    TestCase:
      type: object
      properties:
        id:
          type: integer
        useCaseId:
          type: integer
        caseNumber:
          type: integer
        title:
          type: string
        createdAt:
          type: string
      required: [id, useCaseId, caseNumber, title, createdAt]
...
    CreateTestCaseBody:
      type: object
      properties:
        title:
          type: string
      required: [title]
...
    TestExecution:
      type: object
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        testRunId:
          type: integer
          nullable: true
        iterationNumber:
          type: integer
        testerName:
          type: string
        status:
          type: string
        executedAt:
          type: string
        stepResults:
          type: array
          items:
            $ref: "#/components/schemas/StepResult"
      required: [id, testCaseId, iterationNumber, testerName, status, executedAt]
...
    StepResult:
      type: object
      properties:
        id:
          type: integer
        executionId:
          type: integer
        stepId:
          type: integer
        actualResult:
          type: string
          nullable: true
        comments:
          type: string
          nullable: true
        passed:
          type: boolean
          nullable: true
        recordedAt:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
      required: [id, executionId, stepId, recordedAt, attachments]
...
    Attachment:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        field:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        createdAt:
          type: string
      required: [id, entityType, entityId, field, fileName, fileUrl, fileType, createdAt]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, username, name, role, isActive, createdAt]
...
    Defect:
      type: object
      properties:
        id:
          type: integer
        testRunId:
          type: integer
        testCaseId:
          type: integer
        executionId:
          type: integer
        testerNotes:
          type: string
          nullable: true
        status:
          type: string
        retestReason:
          type: string
          nullable: true
        acceptedByBusinessNote:
          type: string
          nullable: true
        rejectionLog:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        testCase:
          $ref: "#/components/schemas/TestCase"
          nullable: true
        execution:
          $ref: "#/components/schemas/TestExecution"
          nullable: true
        notes:
          type: array
          items:
            $ref: "#/components/schemas/DefectNote"
      required: [id, testRunId, testCaseId, executionId, status, createdAt, updatedAt]
...
    DefectNote:
      type: object
      properties:
        id:
          type: integer
        defectId:
          type: integer
        discussionId:
          type: integer
          nullable: true
        addedByUserId:
          type: integer
        note:
          type: string
        createdAt:
          type: string
          format: date-time
        addedByUser:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, defectId, addedByUserId, note, createdAt]
...
    TeamDiscussion:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        testRunId:
          type: integer
        initiatedByUserId:
          type: integer
        meetingType:
          type: string
          enum: [defect_review, post_mortem]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        participants:
          type: array
          items:
            $ref: "#/components/schemas/DiscussionParticipant"
        initiatedBy:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, projectId, testRunId, initiatedByUserId, meetingType, isActive, createdAt]
    DiscussionParticipant:
      type: object
      properties:
        id:
          type: integer
        discussionId:
          type: integer
        userId:
          type: integer
        canAddNotes:
          type: boolean
        addedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, discussionId, userId, canAddNotes, addedAt]
    CreateDiscussionBody:
      type: object
      properties:
        meetingType:
          type: string
          enum: [defect_review, post_mortem]
        participantIds:
          type: array
          items:
            type: integer
      required: [meetingType, participantIds]
    AddParticipantBody:
      type: object
      properties:
        userId:
          type: integer
        canAddNotes:
          type: boolean
          default: false
      required: [userId]

    # ── Developer Dashboard ─────────────────────────────────────────────────
...